move presentation logic to /views
